Support deserializing HashWithIndifferentAccess
We use serialized columns (via ActiveRecord::Base) heavily in our Rails app,
and most of them are hashes of the HashWithIndifferentAccess variety. Rather
than auditing everywhere a deserialized Hash is used and ensuring that it's
with_indifferent_access, the short term solution is to just have safe_yaml
support it natively.

(This only works with Syck, and is not that clean. A more complete whitelisting
arrangement would be better.)